Art Additions
Notes/Concerns:

- Created a second palette with mostly working autotile walls based on an extension of the base assets.

- Converted some of the frames into animations for players/enemies. Found in Assets/Animations, they are mostly idle/run animations. Attack animations would probably involve attaching a seperate object with its own animation (Using the weapon assets in Assets/Art/frames).

- Created mockup of blind/darkness effect.

Installed the Input System Package and enabled both old/new backends. This can be changed in Edit > Project Settings > Player > Active Input Handling

Made a joystick. The idea is to have one joystick for movement, another joystick for shooting, and miscellaneous buttons for abilities etc. if we are going to make the game playable on mobile.

sphere.png can be used as a projectile texture possibly